+++ br/pkg/backup/schema.go
@@ -21,6 +21,7 @@ import (
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/statistics/handle"
+	"github.com/pingcap/tidb/statistics/handle/storage"
 	kvutil "github.com/tikv/client-go/v2/util"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
@@ -38,7 +39,7 @@ type schemaInfo struct {
 	crc64xor   uint64
 	totalKvs   uint64
 	totalBytes uint64
-	stats      *handle.JSONTable
+	stats      *storage.JSONTable
 }
 
 type iterFuncTp func(kv.Storage, func(*model.DBInfo, *model.TableInfo)) error
