+++ br/pkg/metautil/BUILD.bazel
@@ -11,7 +11,7 @@ go_library(
         "//br/pkg/storage",
         "//br/pkg/summary",
         "//parser/model",
-        "//statistics/handle",
+        "//statistics/handle/storage",
         "//tablecodec",
         "//util/encrypt",
         "@com_github_docker_go_units//:go-units",
