+++ br/pkg/metautil/metafile.go
@@ -24,7 +24,7 @@ import (
 	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/br/pkg/summary"
 	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/statistics/handle"
+	handle "github.com/pingcap/tidb/statistics/handle/storage"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/pingcap/tidb/util/encrypt"
 	"go.uber.org/zap"
