+++ br/pkg/utils/schema_test.go
@@ -14,7 +14,7 @@ import (
 	"github.com/pingcap/tidb/br/pkg/metautil"
 	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/parser/model"
-	"github.com/pingcap/tidb/statistics/handle"
+	handle "github.com/pingcap/tidb/statistics/handle/storage"
 	"github.com/pingcap/tidb/tablecodec"
 	"github.com/stretchr/testify/require"
 )
