+++ executor/BUILD.bazel
@@ -167,6 +167,7 @@ go_library(
         "//statistics",
         "//statistics/handle",
         "//statistics/handle/cache",
+        "//statistics/handle/storage",
         "//store/driver/backoff",
         "//store/driver/error",
         "//store/driver/txn",
