+++ server/handler/optimizor/BUILD.bazel
@@ -19,6 +19,7 @@ go_library(
         "//session",
         "//sessionctx/variable",
         "//statistics/handle",
+        "//statistics/handle/storage",
         "//table",
         "//types",
         "//util",
@@ -57,6 +58,7 @@ go_test(
         "//session",
         "//statistics/handle",
         "//statistics/handle/globalstats",
+        "//statistics/handle/storage",
         "//store/mockstore/unistore",
         "//testkit",
         "//testkit/testsetup",
