+++ statistics/handle/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
         "gc.go",
         "handle.go",
         "handle_hist.go",
-        "historical_stats_handler.go",
         "lock_stats_handler.go",
         "update.go",
     ],
@@ -34,6 +33,7 @@ go_library(
         "//statistics/handle/cache",
         "//statistics/handle/extstats",
         "//statistics/handle/globalstats",
+        "//statistics/handle/history",
         "//statistics/handle/lockstats",
         "//statistics/handle/metrics",
         "//statistics/handle/storage",
@@ -42,15 +42,12 @@ go_library(
         "//types",
         "//util",
         "//util/chunk",
-        "//util/compress",
         "//util/logutil",
         "//util/mathutil",
         "//util/sqlexec",
-        "@com_github_klauspost_compress//gzip",
         "@com_github_ngaut_pools//:pools",
         "@com_github_pingcap_errors//:errors",
         "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_pingcap_tipb//go-tipb",
         "@com_github_tiancaiamao_gp//:gp",
         "@com_github_tikv_client_go_v2//oracle",
         "@org_uber_go_atomic//:atomic",
@@ -83,6 +80,7 @@ go_test(
         "//statistics",
         "//statistics/handle/globalstats",
         "//statistics/handle/internal",
+        "//statistics/handle/storage",
         "//statistics/handle/usage",
         "//testkit",
         "//testkit/testsetup",
