+++ statistics/handle/handle.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"github.com/pingcap/tidb/statistics/handle/history"
 	"math"
 	"strconv"
 	"sync"
@@ -557,7 +558,7 @@ func (h *Handle) SaveTableStatsToStorage(results *statistics.AnalyzeResults, ana
 
 // SaveTableStatsToStorage saves the stats of a table to storage.
 func SaveTableStatsToStorage(sctx sessionctx.Context, results *statistics.AnalyzeResults, analyzeSnapshot bool, source string) error {
-	return storage.SaveTableStatsToStorage(sctx, recordHistoricalStatsMeta, results, analyzeSnapshot, source)
+	return storage.SaveTableStatsToStorage(sctx, history.RecordHistoricalStatsMeta, results, analyzeSnapshot, source)
 }
 
 // SaveStatsToStorage saves the stats to storage.
@@ -577,26 +578,6 @@ func (h *Handle) SaveMetaToStorage(tableID, count, modifyCount int64, source str
 	return h.callWithExec(func(sctx sessionctx.Context, _ sqlexec.RestrictedSQLExecutor) error {
 		return storage.SaveMetaToStorage(sctx, h.recordHistoricalStatsMeta, tableID, count, modifyCount, source)
 	})
-}
-
-func (h *Handle) statsMetaByTableIDFromStorage(tableID int64, snapshot uint64) (version uint64, modifyCount, count int64, err error) {
-	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-	var rows []chunk.Row
-	if snapshot == 0 {
-		rows, _, err = h.execRestrictedSQL(ctx, "SELECT version, modify_count, count from mysql.stats_meta where table_id = %? order by version", tableID)
-	} else {
-		rows, _, err = h.execRestrictedSQLWithSnapshot(ctx, "SELECT version, modify_count, count from mysql.stats_meta where table_id = %? order by version", snapshot, tableID)
-		if err != nil {
-			return 0, 0, 0, err
-		}
-	}
-	if err != nil || len(rows) == 0 {
-		return
-	}
-	version = rows[0].GetUint64(0)
-	modifyCount = rows[0].GetInt64(1)
-	count = rows[0].GetInt64(2)
-	return
 }
 
 func (h *Handle) getGlobalStatsReader(snapshot uint64) (reader *storage.StatsReader, err error) {
@@ -827,7 +808,7 @@ const maxColumnSize = 6 << 20
 // RecordHistoricalStatsToStorage records the given table's stats data to mysql.stats_history
 func (h *Handle) RecordHistoricalStatsToStorage(dbName string, tableInfo *model.TableInfo, physicalID int64, isPartition bool) (uint64, error) {
 	ctx := kv.WithInternalSourceType(context.Background(), kv.InternalTxnStats)
-	var js *JSONTable
+	var js *storage.JSONTable
 	var err error
 	if isPartition {
 		js, err = h.tableStatsToJSON(dbName, tableInfo, physicalID, 0)
@@ -848,7 +829,7 @@ func (h *Handle) RecordHistoricalStatsToStorage(dbName string, tableInfo *model.
 			}
 		}
 	}
-	blocks, err := JSONTableToBlocks(js, maxColumnSize)
+	blocks, err := storage.JSONTableToBlocks(js, maxColumnSize)
 	if err != nil {
 		return version, errors.Trace(err)
 	}
@@ -969,4 +950,41 @@ func getSessionTxnStartTS(se interface{}) (uint64, error) {
 		return 0, err
 	}
 	return txn.StartTS(), nil
+}
+
+const (
+	// StatsMetaHistorySourceAnalyze indicates stats history meta source from analyze
+	StatsMetaHistorySourceAnalyze = "analyze"
+	// StatsMetaHistorySourceLoadStats indicates stats history meta source from load stats
+	StatsMetaHistorySourceLoadStats = "load stats"
+	// StatsMetaHistorySourceFlushStats indicates stats history meta source from flush stats
+	StatsMetaHistorySourceFlushStats = "flush stats"
+	// StatsMetaHistorySourceSchemaChange indicates stats history meta source from schema change
+	StatsMetaHistorySourceSchemaChange = "schema change"
+)
+
+func (h *Handle) recordHistoricalStatsMeta(tableID int64, version uint64, source string) {
+	v := h.statsCache.Load()
+	if v == nil {
+		return
+	}
+	sc := v
+	tbl, ok := sc.Get(tableID)
+	if !ok {
+		return
+	}
+	if !tbl.IsInitialized() {
+		return
+	}
+	err := h.callWithExec(func(sctx sessionctx.Context, _ sqlexec.RestrictedSQLExecutor) error {
+		return history.RecordHistoricalStatsMeta(sctx, tableID, version, source)
+	})
+	if err != nil {
+		logutil.BgLogger().Error("record historical stats meta failed",
+			zap.Int64("table-id", tableID),
+			zap.Uint64("version", version),
+			zap.String("source", source),
+			zap.Error(err))
+		return
+	}
 }
